+++ CHANGELOG.md
@@ -7,6 +7,20 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 [Unreleased changes](./UNRELEASED.md)
 
+# 12.2.0
+
+- Prefer default arguments to default props in functional components
+
+# 12.1.1
+
+## Updated
+
+- Take into account the new Backpack single package in import groups.
+
+# 12.1.0
+
+- Bump eslint-plugin-react from 7.28.0 to 7.30.0
+
 # 12.0.0
 
 ## Breaking